ディレクトリ構成の整理
ディレクトリ構成を、ベストプラクティスに従うよう変更しました。

https://github.com/golang-standards/project-layout

- api ... アプリケーションの基本要素
- cmd ... アプリケーションのエントリーポイント
- internal ... プライベートな package 群
- scripts ... アプリケーションそのものには影響しないコード (スクリプト)
- volume ... DB などのデータ周り
