+++ back/README.md
@@ -4,7 +4,7 @@ golang ([Gin](https://github.com/gin-gonic/gin), [gorm](https://github.com/go-go
 ### API 一覧
 
 ```
-GET    /v1/ping # 疎通確認用
+GET    /ping # 疎通確認用
 GET    /v1/users # ユーザ一覧取得
 GET    /v1/users/:id # ユーザ取得
 PUT    /v1/users/:id # ユーザ更新
@@ -18,7 +18,7 @@ DELETE /v1/users/auth # ユーザ認証無効化
 
 ##### 疎通確認
 ```
-$ curl 0.0.0.0:3000/v1/ping
+$ curl 0.0.0.0:3000/ping
 > pong
 ```
 
@@ -36,6 +36,6 @@ $ curl 0.0.0.0:3000/v1/users/1
 $ curl -X DELETE -H 'Authorization: Bearer <token>' 0.0.0.0:3000/v1/users/1
 
 # 認証
-$ curl -X PUT -H 0.0.0.0:3000/v1/users/1/auth -H 'Content-Type: application/json' -d '{"login": "user", "password": "user1234"}'
+$ curl -X PUT -H 0.0.0.0:3000/v1/users/auth -H 'Content-Type: application/json' -d '{"login": "user", "password": "user1234"}'
 > <token>
 ```
