+++ back/src/api/handler/v1/users.go
@@ -1,12 +1,13 @@
-package api
+package v1
 
 import (
 	"net/http"
 	"strconv"
 
-	"github.com/SongCastle/KoR/lib/jwt"
-	"github.com/SongCastle/KoR/middleware"
-	"github.com/SongCastle/KoR/model"
+	"github.com/SongCastle/KoR/api/middleware"
+	"github.com/SongCastle/KoR/api/model"
+	"github.com/SongCastle/KoR/internal/ecode"
+	"github.com/SongCastle/KoR/internal/jwt"
 	"github.com/gin-gonic/gin"
 )
 
@@ -178,4 +179,12 @@ func UnauthUser(c *gin.Context) {
 
 func responseKeys() []string {
 	return []string{"id", "login", "email"}
+}
+
+func abortWithError(c *gin.Context, status int, code string, err error) {
+	c.AbortWithError(status, err).SetMeta(ecode.CodeJson(code))
+}
+
+func abortWithJSON(c *gin.Context, status int, code string) {
+	c.AbortWithStatusJSON(status, ecode.CodeJson(code))
 }
