+++ back/src/internal/encryptor/encryptor.go
